+++ src/modules/tasks/tasks.service.ts
@@ -3,32 +3,38 @@ import { TCreateTask } from './dto/create-task.dto';
 import { TUpdateTask } from './dto/update-task.dto';
 import { DrizzleAsyncProvider } from 'src/db/drizzle.provider';
 import * as schema from 'src/db/schemas';
-import { eq } from 'drizzle-orm';
+import { and, eq } from 'drizzle-orm';
 import { NeonHttpDatabase } from 'drizzle-orm/neon-http';
+import { MySql2Database } from 'drizzle-orm/mysql2';
 
 @Injectable()
 export class TasksService {
   constructor(
-    @Inject(DrizzleAsyncProvider) private db: NeonHttpDatabase<typeof schema>
+    @Inject(DrizzleAsyncProvider) private db: MySql2Database<typeof schema>
   ) {}
 
-  async create(createTaskDto: TCreateTask) {
-    return await this.db.insert(schema.tasks).values(createTaskDto);
+  async create(userId: number, createTaskDto: TCreateTask) {
+    return await this.db.insert(schema.tasks).values({
+      ...createTaskDto,
+      user_id: userId
+    });
   }
 
-  async findAll() {
-    return await this.db.select().from(schema.tasks);
+  async findAll(userId: number) {
+    return await this.db.select().from(schema.tasks).where(eq(schema.tasks.user_id, userId));
   }
 
-  async findOne(id: number) {
-    return await this.db.select().from(schema.tasks).where(eq(schema.tasks.id, id));
+  async findOne(userId: number, id: number) {
+    return await this.db.query.tasks.findFirst({
+      where: and(eq(schema.tasks.id, id), eq(schema.tasks.user_id, userId))
+    });
   }
 
-  async update(id: number, updateTaskDto: TUpdateTask) {
-    return await this.db.update(schema.tasks).set(updateTaskDto).where(eq(schema.tasks.id, id));
+  async update(userId: number, id: number, updateTaskDto: TUpdateTask) {
+    return await this.db.update(schema.tasks).set(updateTaskDto).where(and(eq(schema.tasks.id, id), eq(schema.tasks.user_id, userId)));
   }
 
-  async remove(id: number) {
-    return await this.db.delete(schema.tasks).where(eq(schema.tasks.id, id));
+  async remove(userId: number, id: number) {
+    return await this.db.delete(schema.tasks).where(and(eq(schema.tasks.id, id), eq(schema.tasks.user_id, userId)));
   }
 }
